requirements.txt
streamlit
tensorflow
pillow
numpy
h5py